backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
email-validator==2.1.0
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
slowapi==0.1.9
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-json-logger==2.0.7
python-dotenv==1.0.0
